Use framework versions that support mapbox 11.3.0

+++ ios/Frameworks/MapboxNavigationNative.xcframework/ios-arm64_x86_64-simulator/MapboxNavigationNative.framework/Headers/MBNNRouteAlternativesOptions.h
@@ -12,11 +12,11 @@ __attribute__((visibility ("default")))
 // This class provides custom init which should be called
 + (nonnull instancetype)new NS_UNAVAILABLE;
 
-- (nonnull instancetype)initWithRequestIntervalSeconds:(uint16_t)requestIntervalSeconds
-                          minTimeBeforeManeuverSeconds:(float)minTimeBeforeManeuverSeconds;
+- (nonnull instancetype)initWithRequestIntervalSeconds:(double)requestIntervalSeconds
+                          minTimeBeforeManeuverSeconds:(double)minTimeBeforeManeuverSeconds;
 
 /** Re-request alternative routes interval. 3 minutes by default. Minimum 30 seconds. */
-@property (nonatomic, readonly) uint16_t requestIntervalSeconds;
+@property (nonatomic, readonly) double requestIntervalSeconds;
 
 /**
  * Allowable travel time to the closest significant maneuvers.
@@ -24,7 +24,7 @@ __attribute__((visibility ("default")))
  *  corresponding to this parameter and the current speed.
  *  0 by default.
  */
-@property (nonatomic, readonly) float minTimeBeforeManeuverSeconds;
+@property (nonatomic, readonly) double minTimeBeforeManeuverSeconds;
 
 
 - (BOOL)isEqualToRouteAlternativesOptions:(nonnull MBNNRouteAlternativesOptions *)other;
